requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
pytest>=7.4.0
python-multipart>=0.0.6
